+++ controllers/indexController.js
@@ -1,9 +1,16 @@
 const asyncHandler = require("express-async-handler");
 const { matchedData, validationResult } = require("express-validator");
 const validatePost = require("../validators/postValidator");
-const { insertPost } = require("../db/queries");
+const { insertPost, getAllPosts, deletePost } = require("../db/queries");
 
 const indexController = {
+  setAllPosts: asyncHandler(async (req, res, next) => {
+    console.log("setAllPosts running...");
+    const posts = await getAllPosts();
+    console.log("posts:", posts);
+    res.locals.posts = posts;
+    next();
+  }),
   getIndex: asyncHandler(async (req, res) => {
     console.log("getIndex running...");
     console.log("req.url", req.url);
@@ -33,6 +40,13 @@ const indexController = {
       res.redirect("/");
     }),
   ],
+  postDeletePost: asyncHandler(async (req, res) => {
+    console.log("postDeletePost running...");
+    const { postID, accountID } = req.params;
+    console.log("req.params:", req.params);
+    await deletePost({ accountID, postID });
+    res.redirect("/");
+  }),
 };
 
 module.exports = indexController;