+++ controllers/profileController.js
@@ -18,6 +18,7 @@ const profileController = {
   }),
   getProfilePosts: asyncHandler(async (req, res) => {
     // Show all posts corresponding to the req.params.username
+    // Do not
     console.log("getProfilePosts running...");
     const { id: accountID } = res.locals.profile;
     const posts = await getAccountPosts({ accountID });